Factored DataSummaryPane and PlotSummaryPane into their own names spaces with their appropriate summarizers.  Exposed these Panes and their namespaces directly in tnt. Cleaned up the FeatureImportanceSummarizer a bit allowing for categorical variables and appropriately rescaling features.
